feat: rename your information component to contact component
