+++ src/pages/CreateAccount/Contact/index.ts
@@ -1,0 +1,1 @@
+export * from './Contact'
