--- src/pages/CreateAccount/YourInformation/index.ts
@@ -1,1 +1,0 @@
-export * from './YourInformation'
